Manifest v3 migration
