+++ lib/WexExtManifestPlugin.js
@@ -4,6 +4,9 @@ const merge = require('deepmerge')
 const open = require('open')
 const argv = require('yargs-parser')(process.argv.slice(2))
 
+const manifestV3Compilation = require('./manifestV3');
+const manifestV2Compilation = require('./manifestV2');
+
 class WexExtManifestPlugin {
   constructor (options, neutrinoOpts) {
     this.options = options
@@ -18,77 +21,6 @@ class WexExtManifestPlugin {
           version: this.neutrinoOpts.packageJson.version
         }
 
-        compilation.entrypoints.forEach((entry, name) => {
-          const htmlPluginOpts = this.neutrinoOpts.mains[name]
-          const entryOpts = htmlPluginOpts.webext
-          if (!entryOpts) {
-            return
-          }
-
-          switch (entryOpts.type) {
-            case 'content_scripts':
-              {
-                if (!neutrinoManifest.content_scripts) {
-                  neutrinoManifest.content_scripts = []
-                }
-
-                const files = entry
-                  .getFiles()
-                  .map(file => file.replace(/\.(css|js)\?.*$/, '.$1'))
-                const js = files.filter(file => file.endsWith('.js'))
-                const css = files.filter(file => file.endsWith('.css'))
-                const opt = {
-                  ...(entryOpts.manifest || {})
-                }
-                if (this.options.polyfill) {
-                  js.unshift('assets/browser-polyfill.min.js')
-                }
-                if (js.length > 0) {
-                  opt.js = opt.js ? [...js, ...opt.js] : js
-                }
-                if (css.length > 0) {
-                  opt.css = opt.css ? [...css, ...opt.css] : css
-                }
-                neutrinoManifest.content_scripts.push(opt)
-              }
-              break
-            case 'background':
-              {
-                const scripts = entry
-                  .getFiles()
-                  .map(file => file.replace(/\.js\?.*$/, '.$1'))
-                  .filter(file => file.endsWith('.js'))
-                if (this.options.polyfill) {
-                  scripts.unshift('assets/browser-polyfill.min.js')
-                }
-                neutrinoManifest[entryOpts.type] = {
-                  scripts,
-                  ...(entryOpts.manifest || {})
-                }
-              }
-              break
-            case 'browser_action':
-            case 'page_action':
-              neutrinoManifest[entryOpts.type] = {
-                default_popup: htmlPluginOpts.filename || `${name}.html`,
-                ...(entryOpts.manifest || {})
-              }
-              break
-            case 'options_page':
-              neutrinoManifest[entryOpts.type] =
-                htmlPluginOpts.filename || `${name}.html`
-              break
-            case 'options_ui':
-              neutrinoManifest[entryOpts.type] = {
-                page: htmlPluginOpts.filename || `${name}.html`,
-                ...(entryOpts.manifest || {})
-              }
-              break
-            default:
-              break
-          }
-        })
-
         let commonManifest = {}
         try {
           commonManifest = require(path.join(
@@ -96,6 +28,16 @@ class WexExtManifestPlugin {
             'common.manifest'
           ))
         } catch(e) {}
+
+        switch (commonManifest.manifest_version) {
+          case 3:
+            manifestV3Compilation(this, compilation, neutrinoManifest)
+            break
+          case 2:
+          default:
+            manifestV2Compilation(this, compilation, neutrinoManifest)
+            break
+        }
 
         const browsers = (await fse.readdir(path.join(this.options.manifest)))
           .map(fileName => (fileName.match(/^([^.]+)\.manifest/) || [, ''])[1])
@@ -208,10 +150,22 @@ function liverreloadModify(manifest) {
   }
   manifest.permissions = [...new Set([
     ...manifest.permissions,
-    "*://neutrino-webextension.reloads/*",
+    manifest.version === 2 && "*://neutrino-webextension.reloads/*",
     "webRequest",
     "webRequestBlocking",
     // remove history
     "browsingData"
   ])]
+
+  if (manifest.version === 3) {
+    if (!manifest.host_permissions) {
+      manifest.host_permissions = []
+    }
+    manifest.host_permissions = [
+      ...new Set([
+        ...manifest.host_permissions,
+        "*://neutrino-webextension.reloads/*"
+      ])
+    ]
+  }
 }
