+++ lib/index.js
@@ -170,7 +170,7 @@ module.exports = (opts = {}) => neutrino => {
         .clear()
         .merge(['browser'])
 
-    neutrino.config.output.globalObject('window')
+    neutrino.config.output.globalObject('this')
 
     // support native dynamic import
     neutrino.config.target(WebExtensionTarget(neutrino.config.node.entries()))
