+++ test/src/manifest/chrome.manifest.json
@@ -1,5 +1,1 @@
-{
-  "background": {
-    "persistent": false
-  }
-}
+{}
