+++ test/src/manifest/common.manifest.json
@@ -1,13 +1,17 @@
 {
-  "manifest_version": 2,
+  "manifest_version": 3,
   "name": "extension_name",
   "short_name": "extension_short_name",
   "description": "extension_description",
-  "web_accessible_resources": [
-    "assets/*"
-  ],
+  "web_accessible_resources": [{
+    "resources": ["assets/*"],
+    "matches": [],
+    "extension_ids": []
+   }],
   "permissions": [
-    "<all_urls>",
     "storage"
+  ],
+  "host_permissions": [
+    "<all_urls>"
   ]
 }
